User content transfer

+++ src/modals/UserDelete.vue
@@ -10,13 +10,19 @@
 				<div class="content">
 					<p>{{ $t('text.reallyDeleteUser', { name: userName }) }}</p>
 					<p>{{ $t('text.cannotBeUndone') }}</p>
+					<p>{{ $t('text.selectUserForTransfer', { name: userName }) }}</p>
+					<select v-model="transferUser" class="form-select">
+						<option v-for="(user, key) in assignableUsers" :value="key">
+							{{ user.name }}
+						</option>
+					</select>
 				</div>
 			</div>
 			<div class="modal-footer">
 				<a class="btn btn-link btn-gray" href="#" aria-label="Cancel" @click.prevent="$emit('closed')">
 					{{ $t('button.cancel') }}
 				</a>
-				<button class="btn btn-error ml-2" @click="deleteUser">{{ $t('button.delete') }}</button>
+				<button class="btn btn-error ml-2" :class="{ disabled: !transferUser }" @click="deleteUser">{{ $t('button.delete') }}</button>
 			</div>
 		</div>
 	</div>
@@ -29,24 +35,52 @@ export default {
 		active: Boolean,
 		userName: String,
 		userKey: String,
-		approved: Boolean
+		approved: Boolean,
+		users: Object,
+		setlists: Object,
+	},
+	data () {
+		return {
+			transferUser: null
+		}
 	},
 	methods: {
 		deleteUser () {
-			// delete approved user (living in users table) or unapproved user (living in registrations table)
-			this.$db.collection(this.approved ? 'users' : 'registrations').doc(this.userKey).delete().then(() => {
-				if (this.approved) {
-					this.$db.collection('permissions').doc(this.userKey).delete();
-				}
-				this.$emit('closed');
-				// toast success message
-				this.$notify({
-					title: this.$parent.$t('toast.userDeleted'),
-					text: this.$parent.$t('toast.userDeletedText'),
-					type: 'primary'
-				});
-			}).catch((error) => this.throwError(error));
-		}
+			if (this.numberOfUsers > 1) {
+				// delete approved user (living in users table) or unapproved user (living in registrations table)
+				this.$db.collection(this.approved ? 'users' : 'registrations').doc(this.userKey).delete().then(() => {
+					if (this.approved) {
+						this.$db.collection('permissions').doc(this.userKey).delete();
+					}
+					// transfer content to selected user
+					for (const setlistKey in this.setlists) {
+						if (Object.hasOwnProperty.call(this.setlists, setlistKey)) {
+							const setlist = this.setlists[setlistKey];
+							if (setlist.creator == this.userKey) {
+								this.$db.collection('setlists').doc(setlistKey).update({ creator: this.transferUser });
+							}
+						}
+					}
+					this.$emit('closed');
+					// toast success message
+					this.$notify({
+						title: this.$parent.$t('toast.userDeleted'),
+						text: this.$parent.$t('toast.userDeletedText'),
+						type: 'primary'
+					});
+				}).catch((error) => this.throwError(error));
+			}
+			}
+	},
+	computed: {
+		assignableUsers () {
+			let users = JSON.parse(JSON.stringify(this.users));
+			delete users[this.userKey];
+			return users;
+		},
+		numberOfUsers () {
+			return Object.keys(this.users).length;
+		},
 	}
 }
 </script>
